Fix: Allgemeine Arbeiten im Backend – isGeneralEntries-Erkennung und PDF-Variable
- intervention_equipment_details.php: GETPOST('equipment_id','int') gibt String
  zurück, nicht Integer. Daher schlägt === 0 (strict) fehl → isGeneralEntries
  blieb false, obwohl equipment_id=0 in der URL. Fix: (int)GETPOST() casten.
- pdf_equipmentmanager.modules.php: Undefinierte Variable $is_first_section
  durch korrekten Ausdruck (count($equipment_list) === 0) ersetzt.

+++ intervention_equipment_details.php
@@ -567,7 +567,7 @@ if ($object->id > 0) {
 
     // Allow page even without equipment (for general entries)
     // equipment_id = 0 means "general entries"
-    $isGeneralEntries = (GETPOSTISSET('equipment_id') && GETPOST('equipment_id', 'int') === 0);
+    $isGeneralEntries = (GETPOSTISSET('equipment_id') && (int)GETPOST('equipment_id', 'int') === 0);
 
     if (count($linked_equipment) == 0 && !$isGeneralEntries) {
         // Show info but also offer "Allgemeine Arbeiten" option
